add test for user sql db
